Dep-205: 댓글 대댓글 도메인 통합 (#72)
* refactor: 대댓글 작성 API 리팩토링

* refactor: 대댓글 조회 API 리팩터링

* refactor: 대댓글 삭제 API 제거 및 댓글 삭제 API 리팩터링

* refactor: 사용하지 않는 코드 제거

+++ Api/src/main/java/com/dingdong/api/community/dto/CommunityIdCardsDto.java
@@ -32,9 +32,9 @@ public class CommunityIdCardsDto {
     private final List<String> keywordTitles;
 
     @Schema(description = "댓글 개수")
-    private final int commentCount;
+    private final Long commentCount;
 
-    public static CommunityIdCardsDto of(IdCard idCard, List<Keyword> keywords, int commentCount) {
+    public static CommunityIdCardsDto of(IdCard idCard, List<Keyword> keywords, Long commentCount) {
         return CommunityIdCardsDto.builder()
                 .idCardId(idCard.getId())
                 .nickname(idCard.getNickname())
